+++ MTrackerUK/Models/Question.swift
@@ -11,5 +11,5 @@ struct Question {
     var id: Int64?
     let assessmentId: Int64
     let questionText: String
-    let questionOrder: Int
+    let questionOrder: Int64
 }
